optimize authentication
