+++ backend/src/controllers/x.controller.ts
@@ -105,7 +105,7 @@ const XCallback = async (req: Request, res: Response) => {
     const option = { httpOnly: true, secure: process.env.NODE_ENV === "develpoment", expires: new Date(Date.now() + (1000 * 60 * 60 * 24)) }
     res.cookie("access_token", userAccessToken, option)
 
-    res.status(200).redirect(`/login?success=${user_details.id}`)
+    res.status(200).redirect(`/login?email=${user_details.email}&uid=${user_details.id}`)
   } catch (error) {
     console.log(error)
     res.status(500).json(error)
